+++ witch_craft/src/core/mod.rs
@@ -1,0 +1,5 @@
+pub mod consts;
+pub mod core;
+pub mod data;
+pub mod structs;
+pub mod types;