+++ witch_craft/src/modules/blackcat/blackcat.rs
@@ -1,0 +1,5 @@
+use crate::{core::types::Closure, modules::blackcat::backend::blackcat_av};
+
+pub fn api() -> Closure {
+    vec![("blackcat.av", blackcat_av)]
+}