+++ witch_craft/src/modules/core/mod.rs
@@ -1,4 +1,0 @@
-pub mod consts;
-pub mod core;
-pub mod data;
-pub mod structs;