+++ witch_craft/src/modules/osint/osint.rs
@@ -1,0 +1,5 @@
+use crate::core::types::Closure;
+
+pub fn api() -> Closure {
+    vec![]
+}