+++ witch_craft/src/test/test_start.rs
@@ -1,6 +1,6 @@
-use crate::modules::core::consts::*;
-use crate::modules::core::core::*;
-use crate::modules::core::structs::DataSet;
+use crate::core::consts::*;
+use crate::core::core::*;
+use crate::core::structs::DataSet;
 use crate::modules::network::structs::*;
 
 #[test]